dom/src/threads/test/test_fibonacci.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads with Fibonacci
-->
<head>
  <title>Test for DOM Worker Threads with Fibonacci</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=450452">DOM Worker Threads Fibonacci</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("fibonacci_worker.js");

  worker.onmessage = function(event) {
    is(event.target, worker);
    is(event.data, 5);
    SimpleTest.finish();
  };

  worker.onerror = function(event) {
    is(event.target, worker);
    ok(false, "Worker had an error: " + event.data);
    SimpleTest.finish();
  };

  worker.postMessage(5);

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
